+++ plugins/destination/test/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [1.3.20](https://github.com/cloudquery/cloudquery/compare/plugins-destination-test-v1.3.19...plugins-destination-test-v1.3.20) (2023-02-14)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.37.1 ([#8008](https://github.com/cloudquery/cloudquery/issues/8008)) ([c47aac0](https://github.com/cloudquery/cloudquery/commit/c47aac0b5e3190a04299713651b97e360043911f))
+
+## [1.3.19](https://github.com/cloudquery/cloudquery/compare/plugins-destination-test-v1.3.18...plugins-destination-test-v1.3.19) (2023-02-14)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.36.0 ([#7809](https://github.com/cloudquery/cloudquery/issues/7809)) ([c85a9cb](https://github.com/cloudquery/cloudquery/commit/c85a9cb697477520e94a1fd260c56b89da62fc87))
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.37.0 ([#7933](https://github.com/cloudquery/cloudquery/issues/7933)) ([dc9cffb](https://github.com/cloudquery/cloudquery/commit/dc9cffbf37bbc6fae73a20bf47e6bbf17e74d1f9))
+
 ## [1.3.18](https://github.com/cloudquery/cloudquery/compare/plugins-destination-test-v1.3.17...plugins-destination-test-v1.3.18) (2023-02-07)